coure_folder_generator/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
idna==3.4
lxml==4.9.2
Pillow==9.5.0
python-docx==0.8.11
pywin32==306
requests==2.31.0
typer==0.9.0
typing_extensions==4.6.3
urllib3==2.0.3
